Quick resetdb variable

+++ frontend/website/__init__.py
@@ -28,9 +28,11 @@ def create_app():
     app.register_blueprint(views, url_prefix="/")
     
     from .models import User, Notification, Follow, Sector, Source, Company, Article
-    with app.app_context():
-        db.drop_all()
-        db.create_all()
+    resetdb = True
+    if resetdb:
+        with app.app_context():
+            db.drop_all()
+            db.create_all()
 
     login_manager = LoginManager()
     login_manager.login_view = 'views.login'
